Accept when remaining table is available
This added test case triggered quite the refactoring.

The deleteFirst-based solution was one I first came across when I did
the exercise in Haskell, where deleteFirstsBy is a standard function in
the base library. I tried to make something more ad hoc, but it wasn't
an improvement on this, so I decided to instead re-implement the
relevant functionality in a Seq module.