Accept reservation by combining two tables
The implementation came accidentally as I was struggling to come up with
something more general. As I was experimenting, all of a sudden, all
tests passed!

I'm not at all happy with the implementation, which I find too
complicated. This is, however, the first time I've done this part of the
kata (in any language), so I'm not sure where this is going. I'll add
more test cases, and hopefully a better implementation will emerge.

I decided to keep the Seq module around for now, although right now,
it's not in use. I may need it later. Othewise, I can delete it.