Reject when group has been reduced
I'm none too pleased with the allocate function, which strikes me as
being too ad hoc. I could probably write it in terms of Seq.fold, but
I'm not sure it'd be easier to understand.

It would also have been easier to write, with no explicit mutation, if
I'd constrained myself to lists, which would have been fine for all
intents and purposes, but now I started down this path of supporting
sequences...